Prevent flashing preview
