+++ src/actions/ast.js
@@ -4,18 +4,9 @@
 
 // @flow
 
-import {
-  getSource,
-  hasSymbols,
-  getSelectedLocation,
-  getSelectedSource,
-  getSelectedFrame,
-  getPreview
-} from "../selectors";
+import { getSource, hasSymbols, getSelectedLocation } from "../selectors";
 
-import { getMappedExpression } from "./expressions";
 import { setInScopeLines } from "./ast/setInScopeLines";
-import { PROMISE } from "./utils/middleware/promise";
 import {
   getSymbols,
   getEmptyLines,
@@ -23,22 +14,8 @@ import {
   isReactComponent
 } from "../workers/parser";
 
-import { findBestMatchExpression } from "../utils/ast";
-
-import { isGeneratedId } from "devtools-source-map";
-
 import type { SourceId } from "debugger-html";
 import type { ThunkArgs } from "./types";
-import type { AstLocation } from "../workers/parser";
-
-const extraProps = {
-  react: { displayName: "this._reactInternalInstance.getName()" },
-  immutable: {
-    isImmutable: exp => `Immutable.Iterable.isIterable(${exp})`,
-    entries: exp => `${exp}.toJS()`,
-    type: exp => `${exp}.constructor.name`
-  }
-};
 
 export function setSourceMetaData(sourceId: SourceId) {
   return async ({ dispatch, getState }: ThunkArgs) => {
@@ -126,126 +103,5 @@ export function setOutOfScopeLocations() {
     });
 
     dispatch(setInScopeLines());
-  };
-}
-
-export function clearPreview() {
-  return ({ dispatch, getState, client }: ThunkArgs) => {
-    const currentSelection = getPreview(getState());
-    if (!currentSelection) {
-      return;
-    }
-
-    return dispatch({
-      type: "CLEAR_SELECTION"
-    });
-  };
-}
-
-export function setPreview(
-  token: string,
-  tokenPos: AstLocation,
-  cursorPos: any
-) {
-  return async ({ dispatch, getState, client, sourceMaps }: ThunkArgs) => {
-    const currentSelection = getPreview(getState());
-    if (currentSelection && currentSelection.updating) {
-      return;
-    }
-
-    await dispatch({
-      type: "SET_PREVIEW",
-      [PROMISE]: (async function() {
-        let immutableType = null;
-        let immutableEntries = null;
-        const source = getSelectedSource(getState());
-        const _symbols = await getSymbols(source.toJS());
-
-        const found = findBestMatchExpression(_symbols, tokenPos, token);
-        if (!found) {
-          return;
-        }
-
-        let { expression, location } = found;
-
-        if (!expression) {
-          return;
-        }
-
-        const sourceId = source.get("id");
-        if (location && !isGeneratedId(sourceId)) {
-          const generatedLocation = await sourceMaps.getGeneratedLocation(
-            { ...location.start, sourceId },
-            source.toJS()
-          );
-
-          expression = await getMappedExpression(
-            { sourceMaps },
-            generatedLocation,
-            expression
-          );
-        }
-
-        const selectedFrame = getSelectedFrame(getState());
-        const { result } = await client.evaluate(expression, {
-          frameId: selectedFrame.id
-        });
-
-        const reactDisplayName = await client.evaluate(
-          extraProps.react.displayName,
-          {
-            frameId: selectedFrame.id
-          }
-        );
-
-        const immutable = await client.evaluate(
-          extraProps.immutable.isImmutable(expression),
-          {
-            frameId: selectedFrame.id
-          }
-        );
-
-        if (immutable.result === true) {
-          immutableEntries = await client.evaluate(
-            extraProps.immutable.entries(expression),
-            {
-              frameId: selectedFrame.id
-            }
-          );
-
-          immutableType = await client.evaluate(
-            extraProps.immutable.type(expression),
-            {
-              frameId: selectedFrame.id
-            }
-          );
-        }
-
-        const extra = {
-          react: {
-            displayName: reactDisplayName.result
-          },
-          immutable: {
-            isImmutable:
-              immutable.result && immutable.result.type !== "undefined",
-            type: immutableType && immutableType.result,
-            entries: immutableEntries && immutableEntries.result
-          }
-        };
-
-        if (result === undefined) {
-          return;
-        }
-
-        return {
-          expression,
-          result,
-          location,
-          tokenPos,
-          cursorPos,
-          extra
-        };
-      })()
-    });
   };
 }
